[Refactor+Fix] Перевод addLikeThunk() под errorObject + ошибки именования add-likes-thunk.ts + забытый экспорт из index.ts. 1. add-likes-thunks.tsx -> add-likes-thunk.ts 2. addLikeThunk() доработан для использования errorObject вместо errorMessage в ветке catch() 3. Из src/thunks/index.ts экспортирован deleteCommentThunk()
